Ajustes-areaComunal: tabla de datos

+++ src/app/admin/Ajustes/ajustes-areas-comunales/ajustes-areas-comunales.component.ts
@@ -1,9 +1,11 @@
-import {Component, OnInit} from '@angular/core';
-import {Router} from '@angular/router';
-import {AreasComunalesService} from '../../../services/areasComunales.service';
-import {Subscription} from "rxjs";
-import {DialogService} from 'src/app/services/dialog.service';
-import {ToastrService} from 'ngx-toastr';
+import { Component, OnInit, ViewChild } from '@angular/core';
+import { Router } from '@angular/router';
+import { AreasComunalesService } from '../../../services/areasComunales.service';
+import { Subscription } from "rxjs";
+import { DialogService } from 'src/app/services/dialog.service';
+import { ToastrService } from 'ngx-toastr';
+import { Query } from '@syncfusion/ej2-data';
+import { CommandModel, GridComponent, PageSettingsModel, PdfExportProperties, ToolbarItems } from '@syncfusion/ej2-angular-grids';
 
 @Component({
   selector: 'app-ajustes-areas-comunales',
@@ -16,12 +18,22 @@ export class AjustesAreasComunalesComponent implements OnInit {
   idCondominio: string = ''
   areasComunales: any[] = [];
 
+  public pageSettings: PageSettingsModel;
+  public toolbarOptions: ToolbarItems[];
+  public commands: CommandModel[];
+  public queryClone: any;
+  @ViewChild('grid') public grid: GridComponent | any;
+
   constructor(
     private router: Router,
     private _areaComunalService: AreasComunalesService,
     private toastr: ToastrService,
     private _dialogService: DialogService
   ) {
+    this.pageSettings = { pageSize: 6 }
+    this.toolbarOptions = ['PdfExport', 'ExcelExport', 'Search'];
+    this.commands = [{ title: 'Editar', buttonOption: { iconCss: 'e-icons e-edit', cssClass: 'e-flat' } },
+    { title: 'Eliminar', buttonOption: { iconCss: 'e-icons e-delete', cssClass: 'e-flat' } }];
     this.recoverData();
   }
 
@@ -47,33 +59,71 @@ export class AjustesAreasComunalesComponent implements OnInit {
     );
   }
 
-  onDelete(id: string) {
+  //Seleccionar editar o eliminar usuario
+  commandClick(item: any): void {
+    if (item.target?.title === 'Editar') {
+      sessionStorage.setItem('idAreaComunal', <string>item.rowData['idAreaComunal']);
+      this.router.navigate(['/admin/ajustes/ajustesAreasComunalesEdit']);
 
-    this._dialogService.confirmDialog({
-      title: 'Eliminar área',
-      message: '¿Está seguro de eliminar el área?',
-      confirmText: 'Si',
-      cancelText: 'No',
-    }).subscribe(res => {
-      if (res) {
-        this._areaComunalService.deleteAreasComunales(id).then(() => {
-          this.toastr.success('El registro fue eliminado con exito', 'Registro eliminado', {
-            positionClass: 'toast-bottom-right'
-          });
-        }).catch(error => {
-          console.log(error);
-        })
-      }
-    });
-
+    } else if (item.target?.title === 'Eliminar') {
+      const id = <string>item.rowData['idAreaComunal'];
+      this._dialogService.confirmDialog({
+        title: 'Eliminar área',
+        message: '¿Está seguro de eliminar el área?',
+        confirmText: 'Si',
+        cancelText: 'No',
+      }).subscribe(res => {
+        if (res) {
+          this._areaComunalService.deleteAreasComunales(id).then(() => {
+            this.toastr.success('El registro fue eliminado con exito', 'Registro eliminado', {
+              positionClass: 'toast-bottom-right'
+            });
+          }).catch(error => {
+            console.log(error);
+          })
+        }
+      });
+    }
   }
 
   onGoCreate() {
     this.router.navigate(['/admin/ajustes/ajustesAreasComunalesCreate']);
   }
 
-  onGoEdit(item: any) {
-    sessionStorage.setItem('idAreaComunal', <string>item.idAreaComunal);
-    this.router.navigate(['/admin/ajustes/ajustesAreasComunalesEdit']);
+  //Seleccionar exportar excel y pdf
+  toolbarClick(args: any): void {
+    if (args.item.id === 'Grid_pdfexport') {
+      const pdfExportProperties: PdfExportProperties = {
+        fileName: 'usuarios.pdf'
+      };
+      this.queryClone = this.grid.query;
+      this.grid.query = new Query().addParams('recordcount', '12');
+      this.grid.pdfExport(pdfExportProperties);
+      //this.grid.pdfExport();
+    } else if (args.item.id === 'Grid_excelexport') {
+      this.queryClone = this.grid.query;
+      this.grid.query = new Query().addParams('recordcount', '12');
+      this.grid.excelExport();
+    }
+  }
+
+  pdfExportComplete(): void {
+    this.grid.query = this.queryClone;
+  }
+
+  excelExportComplete(): void {
+    this.grid.query = this.queryClone;
+  }
+
+  rowDataBound(args: any) {
+    // aquí estamos calculando el número de serie
+    var rowIndex = parseInt(args.row.getAttribute('aria-rowIndex'));
+    var page = this.grid.pageSettings.currentPage! - 1;
+
+    var totalPages = this.grid.pageSettings.pageSize;
+    var startIndex = page * totalPages!;
+    var sno = startIndex + (rowIndex + 1);
+    //  actualizando el valor en la primera celda de la fila donde hemos representado una columna vacía para esto
+    args.row.cells[0].innerText = sno;
   }
 }
